refactor(Web_Ui): 重构日志处理和组件导出逻辑
- 添加自定义日志过滤器以屏蔽开发服务器警告
- 重构组件导出结果处理逻辑，简化返回数据结构
- 将几何计算和符号库工具函数拆分到独立模块
- 优化前端结果展示逻辑
